+++ fractal_server/app/routes/api/v2/task.py
@@ -116,6 +116,22 @@ async def patch_task(
     )
     update = task_update.model_dump(exclude_unset=True)
 
+    task_group = await db.get(TaskGroupV2, db_task.taskgroupv2_id)
+    if task_group.origin != TaskGroupOriginEnum.OTHER:
+        for argname in (
+            "command_parallel",
+            "command_non_parallel",
+            "input_types",
+            "output_types",
+        ):
+            if argname in update:
+                raise HTTPException(
+                    status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
+                    detail=(
+                        f"Cannot update '{argname}' when {task_group.origin=}."
+                    ),
+                )
+
     # Forbid changes that set a previously unset command
     if db_task.type == TaskType.NON_PARALLEL and "command_parallel" in update:
         raise HTTPException(