+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 **Note**: Numbers like (\#1234) point to closed Pull Requests on the fractal-server repository.
 
-# Unreleased
+# 2.20.4
+
+* API:
+    * Forbid editing some individual task properties if the origin of the task group is not "other" (\#3247).
+* Testing:
+    * Update testing database to version 2.20.3 adding workflow templates (\#3253).
+
+# 2.20.3
+
+* API:
+    * Always include `/auth/token/logout/`, independently on `FRACTAL_DISABLE_BASIC_AUTH` (\#3252).
+
+# 2.20.2
 
 * API:
     * Drop `POST /auth/login/` and `POST /auth/logout/` cookie-based endpoints (\#3242, \#3249).